Fix: `VKTexturePool` incorrect format modifier in log output
Quick fix for incorrect format modifiers in `CLOG_TRACE` in
the debug output of `VKTexturePool` giving warnings on clang.

Pull Request: https://projects.blender.org/blender/blender/pulls/153773

+++ source/blender/gpu/vulkan/vk_texture_pool.cc
@@ -374,11 +374,11 @@ void VKTexturePool::log_usage_data()
                 static_cast<float>(total_allocation_size);
 
   CLOG_TRACE(&LOG,
-             "VKTexturePool uses %zu/%zu mb (%.1f%% of %li allocations)",
-             current_usage_data_.acquired_segment_size_max >> 20,
-             total_allocation_size >> 20,
+             "VKTexturePool uses %lu/%lu mb (%.1f%% of %lu allocations)",
+             static_cast<unsigned long>(current_usage_data_.acquired_segment_size_max >> 20),
+             static_cast<unsigned long>(total_allocation_size >> 20),
              ratio * 100.0f,
-             current_usage_data_.allocation_count);
+             static_cast<unsigned long>(current_usage_data_.allocation_count));
 }
 
 }  // namespace blender::gpu
